CBT_Sekolah/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
pandas
openpyxl
python-dotenv
werkzeug